[PH-213] AdArea.tsx 생성, AdPage 정리
